requirements.txt
crewai==0.130.0
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9
openai
PyPDF2==3.0.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
jinja2